[REFACTOR] created placeholder and componentized search_view
